Update README.md
Examined the elements of the profession and the essential skills to excel in a business analysis role.
Developed a comprehensive understanding of the scope of the business analysis profession.
Assisted in identifying a core set of competencies crucial for successful business analysis. This experience has enabled me to establish a connection between business analysis and research.